requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML - Ollama (Local LLM)
# No package needed - Ollama runs separately
# Download from: https://ollama.ai

# Speech Recognition
SpeechRecognition==3.10.1
pyaudio==0.2.14
# Alternative offline STT
# openai-whisper==20231117

# Text-to-Speech
pyttsx3==2.90
gtts==2.5.0

# Wake Word Detection
pvporcupine==3.0.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# GUI (Desktop)
PyQt6==6.6.1
# Alternative: PySide6==6.6.1

# System Integration
pyautogui==0.9.54
psutil==5.9.8
schedule==1.2.1

# Web Search
googlesearch-python==1.2.3
beautifulsoup4==4.12.3

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
